refactor config package
