+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/config/ProxyConfiguration.java
@@ -15,31 +15,23 @@
  * </p>
  */
 
-package io.shardingsphere.shardingproxy.config.yaml;
+package io.shardingsphere.shardingproxy.config;
 
-import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.core.yaml.masterslave.YamlMasterSlaveRuleConfiguration;
-import io.shardingsphere.core.yaml.sharding.YamlShardingRuleConfiguration;
 import lombok.Getter;
-import lombok.Setter;
+import lombok.RequiredArgsConstructor;
 
-import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Rule configuration for yaml.
+ * Proxy configuration.
  *
- * @author panjuan
+ * @author zhangliang
  */
+@RequiredArgsConstructor
 @Getter
-@Setter
-public final class ProxyYamlRuleConfiguration {
+public final class ProxyConfiguration {
     
-    private String schemaName;
+    private final ProxyYamlServerConfiguration serverConfiguration;
     
-    private Map<String, DataSourceParameter> dataSources = new HashMap<>();
-    
-    private YamlShardingRuleConfiguration shardingRule;
-    
-    private YamlMasterSlaveRuleConfiguration masterSlaveRule;
+    private final Map<String, ProxyYamlRuleConfiguration> ruleConfigurationMap;
 }
